New path for community api
Add api/v1 as prefix and update all tests
