+++ seta-ui/seta_flask_server/blueprints/communities/__init__.py
@@ -2,8 +2,9 @@ from flask_restx import Api
 from .community import communities_ns
 from .community_membership import membership_ns
 from .community_change_requests import community_change_request_ns
-from .invites import invite_ns
 from .community_invite import community_invite_ns
+from .community_resources import community_resources_ns
+from .invites import invite_ns
 from .resource import resources_ns
 from .resource_contributor import resource_contributors_ns
 from .resource_change_request import resource_change_request_ns
@@ -28,7 +29,7 @@ api = Api(
          title='SeTA Communities API',
          version='1.0',
          description='SeTA Communities API',
-         doc='/doc',
+         doc='/communities/doc',
          authorizations=authorizations
          )
 
@@ -36,8 +37,12 @@ api.add_namespace(communities_ns, path="/communities")
 api.add_namespace(community_change_request_ns, path="/communities")
 api.add_namespace(membership_ns, path="/communities")
 api.add_namespace(community_invite_ns, path="/communities")
+api.add_namespace(community_resources_ns, path="/communities")
+
 api.add_namespace(invite_ns, path="/invites")
+
 api.add_namespace(resources_ns, path="/resources")
 api.add_namespace(resource_contributors_ns, path="/resources")
 api.add_namespace(resource_change_request_ns, path="/resources")
+
 api.add_namespace(permissions_ns, path="/permissions")
