+++ seta-ui/tests/infrastructure/helpers/community_permission.py
@@ -1,13 +1,15 @@
 from .util import auth_headers
 from flask.testing import FlaskClient
 
+API_V1 = "/api/v1"
+
 def get_user_permissions(client: FlaskClient, access_token:str, community_id: str, user_id: str):
-    url = f"/api/communities/v1/permissions/community/{community_id}/user/{user_id}"
+    url = f"{API_V1}/permissions/community/{community_id}/user/{user_id}"
 
     return client.get(url, headers=auth_headers(access_token))
 
 def replace_user_permissions(client: FlaskClient, access_token:str, community_id: str, user_id: str, scopes:list[str]):
-    url = f"/api/communities/v1/permissions/community/{community_id}/user/{user_id}"
+    url = f"{API_V1}/permissions/community/{community_id}/user/{user_id}"
 
     data =""
     for scope in scopes:
